+++ brewtils/models.py
@@ -129,6 +129,7 @@ class Command(BaseModel):
     def __init__(
         self,
         name=None,
+        display_name=None,
         description=None,
         parameters=None,
         command_type=None,
@@ -144,6 +145,7 @@ class Command(BaseModel):
         allow_any_kwargs=None,
     ):
         self.name = name
+        self.display_name = display_name or name
         self.description = description
         self.parameters = parameters or []
         self.command_type = command_type
@@ -640,6 +642,7 @@ class RequestTemplate(BaseModel):
         "instance_name",
         "namespace",
         "command",
+        "command_display_name",
         "command_type",
         "parameters",
         "comment",
@@ -654,6 +657,7 @@ class RequestTemplate(BaseModel):
         instance_name=None,
         namespace=None,
         command=None,
+        command_display_name=None,
         command_type=None,
         parameters=None,
         comment=None,
@@ -665,6 +669,7 @@ class RequestTemplate(BaseModel):
         self.instance_name = instance_name
         self.namespace = namespace
         self.command = command
+        self.command_display_name = command_display_name or command
         self.command_type = command_type
         self.parameters = parameters
         self.comment = comment
@@ -711,6 +716,7 @@ class Request(RequestTemplate):
         instance_name=None,
         namespace=None,
         command=None,
+        command_display_name=None,
         id=None,  # noqa # shadows built-in
         is_event=None,
         parent=None,
@@ -738,6 +744,7 @@ class Request(RequestTemplate):
             instance_name=instance_name,
             namespace=namespace,
             command=command,
+            command_display_name=command_display_name,
             command_type=command_type,
             parameters=parameters,
             comment=comment,