+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Brewtils Changelog
 TBD
 
 - Update Parameter and Resolvable to add file_name
+- Added support for display name to command decorator
 - Updated Wait Timeout Exception expected HTTP code from 408 to 504
 - Dropping Official Python 2.7 Support